Moved gridfs_patch from ccpncdb/ to tests/py/
