+++ tests/py/server_test.py
@@ -47,7 +47,7 @@ class MDBServerTest(unittest.TestCase):
     @mongomock.patch("mongodb://localhost:27017", on_new="pymongo")
     def setUp(self):
         
-        from ccpncdb.gridfs_patch import enable_gridfs_integration
+        from tests.py.gridfs_patch import enable_gridfs_integration
         enable_gridfs_integration()
 
         from ccpncdb.server import MainServer
